+++ lib/churn_vs_complexity/hotspots/checker.rb
@@ -3,14 +3,15 @@
 module ChurnVsComplexity
   module Hotspots
     class Checker
-      def initialize(engine:, serializer:)
+      def initialize(engine:, serializer:, language: nil)
         @engine = engine
         @serializer = serializer
+        @language = language
       end
 
       def check(folder:)
         raw_result = @engine.check(folder:)
-        @serializer.serialize(raw_result)
+        @serializer.serialize(raw_result.merge(language: @language))
       end
     end
   end